requirements.txt
uproot3
#numpy
pandas
matplotlib
scikit-spatial
tqdm

torch==1.13.0
numba
torch_scatter

psycopg2-binary